+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [UICIRCLOG-34](https://issues.folio.org/browse/UICIRCLOG-34) Make text filters submit on a button or Enter.
 * [UICIRCLOG-28](https://issues.folio.org/browse/UICIRCLOG-28) Auto-focus first search field.
+* [UICIRCLOG-35](https://issues.folio.org/browse/UICIRCLOG-35) Reformat dates to match date format across FOLIO.
 
 ## [1.0.1](https://github.com/folio-org/ui-circulation-log/tree/v1.0.1) (2020-10-30)
 [Full Changelog](https://github.com/folio-org/ui-circulation-log/compare/v1.0.0...v1.0.1)